テストを統合，javaファイルのファイル名を修正
